* Implemented all the countdowns for the generators (diamond, emerald) * Implemented countdown for the crystals dying, and finally, the end of the game (shutdown) * Attempted to correct some implementation of the Countdown (check comments for questions I raised)
Constants were added in PlayerManager for the upgrade times, but converting to a configuration system should be simple by using the DiamondUpgrade#getLevel & EmeraldUpgrade#getLevel.

The limit is also adjustable in the code (the maximum level the generators can upgrade to)

The generators are a type of countdown to avoid code duplication.